\name{apl-package}
\alias{apl-package}
\alias{apl}
\docType{package} 
\title{
APL functions in R
}
\description{
APL functions in R.
}
\details{
\tabular{ll}{
Package: \tab apl\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-09-21\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
APL functions in R.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu}

Maintainer: Masanao Yajima<yajima@stat.ucla.edu>
}
\references{
G. Helzer. (1989):
\emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
