% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl.R
\name{apl_plotly}
\alias{apl_plotly}
\title{Plot Association Plot with plotly}
\usage{
apl_plotly(
  rows,
  rows_group = NULL,
  cols,
  cols_group,
  rows_scored = NULL,
  rows_color = "#0066FF",
  rows_high_color = "#FF0000",
  cols_color = "#601A4A",
  cols_high_color = "#EE442F",
  score_color = "rainbow",
  row_labs = FALSE,
  col_labs = FALSE,
  show_score = FALSE,
  show_cols = FALSE,
  show_rows = TRUE
)
}
\arguments{
\item{rows}{Row APL-coordinates}

\item{rows_group}{Row AP-coordinates to highlight}

\item{cols}{Column AP-coordinates}

\item{cols_group}{Column AP-coordinates for the group to be highlighted.}

\item{rows_scored}{Row AP-coordinates of rows above a score cutoff.}

\item{rows_color}{Color for rows}

\item{rows_high_color}{Color for rows to be highlighted.}

\item{cols_color}{Column points color.}

\item{cols_high_color}{Color for column points to be highlighted.}

\item{score_color}{Color scheme for row points with a score.}

\item{row_labs}{Logical. Whether labels for rows indicated by rows_idx
should be labeled with text. Default TRUE.}

\item{col_labs}{Logical. Whether labels for columns indicated by cols_idx
shouls be labeled with text. Default FALSE.}

\item{show_score}{Logical. Whether the S-alpha score should be shown in
the plot.}

\item{show_cols}{Logical. Whether column points should be plotted.}

\item{show_rows}{Logical. Whether row points should be plotted.}
}
\value{
Interactive plotly Association Plot
}
\description{
Uses plotly to generate an interactive Association Plot
}
