% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl.R
\name{apl_topGO}
\alias{apl_topGO}
\title{Run Gene overrepresentation analysis with topGO}
\usage{
apl_topGO(
  caobj,
  ontology,
  organism = "hs",
  ngenes = 1000,
  top_res = 10,
  use_coords = FALSE,
  return_plot = FALSE
)
}
\arguments{
\item{caobj}{A "cacomp" object with principal row coordinates and standard column coordinates calculated.}

\item{ontology}{Character string. Chooses GO sets for 'BP' (biological processes), 'CC' (cell compartment) or 'MF' (molecular function).}

\item{organism}{Character string. Either 'hs' (homo sapiens), 'mm' (mus musculus) or the name of the organism package such as 'org.*.eg.db'.}

\item{ngenes}{Numeric. Number of top ranked genes to test for overrepresentation.}

\item{top_res}{Numeric. Number of most significant gene sets to return.}

\item{use_coords}{Logical. Whether the x-coordinates of the row APL coordinates should be used for ranking.
Only recommended when no S-alpha score (see apl_score()) can be calculated.}

\item{return_plot}{Logical. Whether a plot of significant gene sets should be additionally returned.}
}
\value{
A data.frame containing the gene sets with the highest overrepresentation.
}
\description{
This function uses the Kolmogorov-Smirnov test as implemented by the package topGO to test for overrepresentation in Gene Ontology gene sets.
}
\details{

}
