\name{aplJoin}
\alias{aplJoin}
\title{
APL Join
}
\description{
Joins 2 arrays together along specified dimension.
}
\usage{
aplJoin(a, b, k)
}
\arguments{
  \item{a}{an array}
  \item{b}{an array}
  \item{k}{a number specifying dimension to join a and b along.}
}
\details{
Array a and b are joined along the dimension k.  
If a is n by m array and b is k by j array to join a and b along 1st dimension, m must be same as j.  Similary to join along 2nd dimension, n must be same as k.
}
\references{
G. Helzer. (1989):
\emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu}
}

\examples{
a<- array(1:6,c(2,3))
b<- array(1:4,c(2,2))
aplJoin(a,b,2)
}
\keyword{ manip }
