\name{aplMemberOf}
\alias{aplMemberOf}
\title{
APL MemberOf
}
\description{
Checks the membership of a in b
}
\usage{
aplMemberOf(a, b)
}
\arguments{
  \item{a}{an array or a matrix that gets checked whether each of its member is in b.}
  \item{b}{a reference array or matrix to check the membeship of a.}
}
\details{
Each member of a is checked if it exists in b.  Result is an object with same dimension as a with each element replaced by 1 or 0, corresponding to if the member at that location in original matrix exists (1) or not (0).
}
\references{
G. Helzer. (1989):
\emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu}
}

\examples{
aplMemberOf(matrix(1:6,3),array(1:4,c(1,4)))
}
\keyword{ manip }
