\name{aplRank}
\alias{aplRank}
\title{
APL Rank
}
\description{
Returns the dimension of an array.
}
\usage{
aplRank(a)
}
\arguments{
  \item{a}{an array}
}
\details{
This function returns the dimension of an array.
}
\references{
G. Helzer. (1989):
\emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu}
}

\seealso{
    \code{\link[apl]{aplShape}}
}
\examples{
    a <- array( 1:10, c(2,5))
    aplRank( a )
    b <- array( 1:27, c(3,3,3))
    aplRank( b )
}
\keyword{ manip }
