\name{aplRavel}
\alias{aplRavel}
\title{
APL Ravel
}
\description{
Reshapes an array or a matrix into a vector
}
\usage{
aplRavel(a)
}
\arguments{
\item{a}{An array or a matrix.}
}
\details{
This function takes an array or a matrix and returns a vector.  
The order of the elements will follow default R rule of column major order.
}
\references{
G. Helzer. (1989):
\emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu}
}
\examples{
aplRavel(array(1:27,c(3,9)))
aplRavel(matrix(1:27,9))
}
\keyword{ manip }
