\name{aplReshape}
\alias{aplReshape}
\title{
APL Reshape
}
\description{
Reshape a vector, an array, or a matrix.
}
\usage{
aplReshape(a, d)
}
\arguments{
  \item{a}{ A vector, an array, or a matrix. }
  \item{d}{ A vector of dimension size to reshape a. }
}
\details{
Reshape will convert $a$ into an array with dimension specified by $d$.
}
\references{
G. Helzer. (1989):
\emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu},
Masanao Yajima \email{yajima@stat.ucla.edu}
}
\seealso{
    \code{\link[apl]{aplShape}}, 
}
\examples{
aplReshape(1:10,c(2,5))
aplReshape(array(1:27,c(3,9)),c(9,3))
aplReshape(matrix(1:27,9),c(3,9))
}
\keyword{ manip }
