\name{aplRTV}
\alias{aplRTV}
\title{
APL Rotate
}
\description{
Shifts the elements of a vector or array dimension.}
\usage{
aplRTV(a, k)
}
\arguments{
  \item{a}{a vector or an array}
  \item{k}{number to shift over}
}
\details{
Shifts the elements of a vector or array dimension.
}

\references{
G. Helzer. (1989):
\emph{An Encyclopedia of APL}, p.191-193, second edition, I-APL LTD, St. Albans, G.B.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu}
}

\seealso{

}
\examples{
}
\keyword{ manip }
