% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.cacomp,cacomp-method}
\alias{as.cacomp,cacomp-method}
\title{Convert cacomp object to cacomp object (returns input).}
\usage{
\S4method{as.cacomp}{cacomp}(obj, assay = NULL, ...)
}
\description{
Convert cacomp object to cacomp object (returns input).
}
