% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.cacomp,Seurat-method}
\alias{as.cacomp,Seurat-method}
\title{Convert Seurat object to cacomp object.}
\usage{
\S4method{as.cacomp}{Seurat}(obj, assay = "RNA", ..., slot = "counts")
}
\description{
Convert Seurat object to cacomp object.
}
