% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.cacomp,SingleCellExperiment-method}
\alias{as.cacomp,SingleCellExperiment-method}
\title{Convert SingleCellObject object to cacomp object.}
\usage{
\S4method{as.cacomp}{SingleCellExperiment}(obj, assay = "counts", ...)
}
\description{
Convert SingleCellObject object to cacomp object.
}
