% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.cacomp.cacomp}
\alias{as.cacomp.cacomp}
\title{Convert to cacomp object}
\usage{
\method{as.cacomp}{cacomp}(obj)
}
\arguments{
\item{obj}{An object of class "cacomp".}
}
\value{
A cacomp object.
}
\description{
If an object already is of class cacomp, the function simply returns the input.
}
