% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.cacomp}
\alias{as.cacomp}
\title{Create cacomp object from Seurat container}
\usage{
as.cacomp(obj, assay, recompute = TRUE)
}
\arguments{
\item{obj}{An object of class "cacomp", "Seurat" or "SingleCellExperiment" with a DimReduc Object in the "CA" slot.}

\item{assay}{The name of the assay to use, e.g. "RNA".}

\item{recompute}{Should all values be recomputed (std_coords_rows, U and V as well as row and column masses)? Default TRUE.}
}
\value{
A cacomp object.
}
\description{
Converts the values stored in the Seurat DimReduc slot "CA" to an cacomp object.
If recompute = TRUE additional parameters are recomputed from the saved values without rerunning SVD.
}
