% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.cacomp.SingleCellExperiment}
\alias{as.cacomp.SingleCellExperiment}
\title{Create cacomp object from SingleCellExperiment object.}
\usage{
\method{as.cacomp}{SingleCellExperiment}(obj, assay = NULL, recompute = TRUE)
}
\arguments{
\item{obj}{An object of class "SingleCellExperiment" with a LinearEmbeddingMatrix object in the reducedDim(sce, "CA") slot.}

\item{assay}{The name of the assay to use, e.g. "RNA".}

\item{recompute}{Should all values be recomputed (std_coords_rows, U and V as well as row and column masses)? Default TRUE.}
}
\value{
A cacomp object.
}
\description{
Converts the values stored in the SingleCellExperiment reducedDim slot "CA" to an cacomp object.
If recompute = TRUE additional parameters are recomputed from the saved values without rerunning SVD.
}
