\name{butLast}
\alias{butLast}
\title{
Everything but the last element
}
\description{
Returns everything but the last element in a vector.}
\usage{
butLast(x)
}
\arguments{
  \item{x}{a vector}
}
\details{
This function takes a vector and returns the original vector without the last element.
}

\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu}
}

\seealso{
    \code{\link[apl]{first}}, 
    \code{\link[apl]{last}}, 
    \code{\link[apl]{butFirst}}
}
\examples{
    x <- 1:10
    butLast(x)
}
\keyword{ manip }

