% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ca_biplot.cacomp}
\alias{ca_biplot.cacomp}
\title{Plot of the first 2 CA dimensions.}
\usage{
\method{ca_biplot}{cacomp}(
  obj,
  xdim = 1,
  ydim = 2,
  princ_coords = 1,
  row_labels = NULL,
  col_labels = seq(ncol(obj$std_coords_cols)),
  type = "plotly"
)
}
\arguments{
\item{obj}{An object of class "cacomp" with standard and principal coordinates calculated.}

\item{xdim}{Integer. The dimension for the x-axis. Default 1.}

\item{ydim}{Integer. The dimension for the y-axis. Default 2.}

\item{princ_coords}{Integer. If 1 then principal coordinates are used for the rows, if 2 for the columns. Default 1 (rows).}

\item{row_labels}{Numeric vector. Indices for the rows for which a label should be added (label should be stored in rownames). Default NULL.}

\item{col_labels}{Numeric vector. Indices for the columns for which a label should be added (label should be stored in colnames).
Default seq(ncol(obj$std_coords_cols)) (all columns).}

\item{type}{String. Type of plot to draw. Either "ggplot" or "plotly". Default "plotly".}
}
\value{
Plot of class "plotly" or "ggplot".
}
\description{
Plots the first 2 dimensions of the rows and columns in the same plot.
}
\details{
Choosing type "plotly" will generate an interactive html plot with the package plotly. Type "ggplot" generates a static plot.
Depending on whether `princ_coords` is set to 1 or 2 either the principal coordinates of either the rows (1) or the columns (2)
are chosen. For the other the standard coordinates are plotted (assymetric biplot).
Labels for rows and columns should be stored in the row and column names respectively.
}
