% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{cacomp}
\alias{cacomp}
\title{Correspondance Analysis}
\usage{
cacomp(
  obj,
  assay,
  coords = TRUE,
  princ_coords = 1,
  python = TRUE,
  dims = min(nrow(S), ncol(S)),
  top = nrow(obj),
  inertia = TRUE,
  rm_zeros = TRUE,
  return_input = FALSE,
  ...
)
}
\arguments{
\item{obj}{A numeric matrix or Seurat/SingleCellExperiment object. For sequencing a count matrix, gene expression values with genes in rows and samples/cells in columns.
Should contain row and column names.}

\item{coords}{Logical. Indicates whether CA standard coordinates should be calculated. Default TRUE}

\item{princ_coords}{Integer. Number indicating whether principal coordinates should be calculated for the rows (=1), columns (=2), both (=3) or none (=0).
Default 1.}

\item{python}{A logical value indicating whether to use singular-value decomposition from the python package torch.
This implementation dramatically speeds up computation compared to `svd()` in R.}

\item{dims}{Integer. Number of CA dimensions to retain. Default NULL (keeps all dimensions).}

\item{top}{Integer. Number of most variable rows to retain. Default NULL.}

\item{inertia}{Logical.. Whether total, row and column inertias should be calculated and returned. Default TRUE.}

\item{rm_zeros}{Logical. Whether rows containing only 0s should be removed. Keeping zero only rows might lead to unexpected results.Default TRUE.}
}
\value{
Returns a named list of class "cacomp" with components
U, V and D: The results from the SVD.
row_masses and col_masses: Row and columns masses.
top_rows: How many of the most variable rows were retained for the analysis.
tot_inertia, row_inertia and col_inertia: Only if inertia = TRUE. Total, row and column inertia respectively.
}
\description{
`cacomp` performs correspondence analysis on a matrix and returns the transformed data.
}
\details{
The calculation is performed according to Greenacre. Singular value decomposition
can be performed either with the base R function `svd` or preferably by the much faster
pytorch implementation (python = TRUE). When working on large matrices, CA coordinates and
principal coordinates should only computed when needed to save computational time.
}
\references{
Greenacre, M. Correspondence Analysis in Practice, Third Edition, 2017.
}
