% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{cacomp.Seurat}
\alias{cacomp.Seurat}
\title{Correspondance Analysis for Seurat objects}
\usage{
\method{cacomp}{Seurat}(
  obj,
  assay = DefaultAssay(obj),
  coords = TRUE,
  princ_coords = 1,
  python = TRUE,
  dims = NULL,
  top = NULL,
  inertia = TRUE,
  rm_zeros = TRUE,
  return_input = FALSE,
  ...
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{assay}{Character string. The assay from which extract the count matrix for SVD. Default DefaultAssay of obj.}

\item{coords}{Logical. Indicates whether CA standard coordinates should be calculated. Default TRUE}

\item{princ_coords}{Integer. Number indicating whether principal coordinates should be calculated for the rows (=1), columns (=2), both (=3) or none (=0).
Default 1.}

\item{python}{A logical value indicating whether to use singular-value decomposition from the python package torch.
This implementation dramatically speeds up computation compared to `svd()` in R.}

\item{dims}{Integer. Number of CA dimensions to retain. Default NULL (keeps all dimensions).}

\item{top}{Integer. Number of most variable rows to retain. Default NULL.}

\item{inertia}{Logical.. Whether total, row and column inertias should be calculated and returned. Default TRUE.}

\item{rm_zeros}{Logical. Whether rows containing only 0s should be removed. Keeping zero only rows might lead to unexpected results.Default TRUE.}

\item{return_input}{Logical. If TRUE returns the input (Seurat object) with the CA saved in the DimReduc slot "CA". Otherwise returns a "cacomp". Default FALSE.}
}
\value{
Returns seurat_obj with a new Dimensional Reduction Object named "CA".
Standard coordinates of the cells are saved as embeddings,
the principal coordinates of the genes as loadings and
the singular values (= square root of principal intertias/eigenvalues)
are stored as stdev.
To recompute a regular "cacomp" object without rerunning cacomp use `ca_from_seurat()`.
}
\description{
`cacomp.seurat` performs correspondence analysis on a matrix and stores the standard coordinates
 of the columns (= cells) and the principal coordinates of the rows (= genes) as a DimReduc Object in the Seurat container.
}
\details{
The calculation is performed according to Greenacre. Singular value decomposition
can be performed either with the base R function `svd` or preferably by the much faster
pytorch implementation (python = TRUE). When working on large matrices, CA coordinates and
principal coordinates should only computed when needed to save computational time.
}
\references{
Greenacre, M. Correspondence Analysis in Practice, Third Edition, 2017.
}
