% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{cacomp.SingleCellExperiment}
\alias{cacomp.SingleCellExperiment}
\title{Correspondance Analysis for SingleCellExperiment objects}
\usage{
\method{cacomp}{SingleCellExperiment}(
  obj,
  assay = "counts",
  coords = TRUE,
  princ_coords = 1,
  python = TRUE,
  dims = NULL,
  top = NULL,
  inertia = TRUE,
  rm_zeros = TRUE,
  return_input = FALSE
)
}
\arguments{
\item{obj}{A SingleCellExperiment object.}

\item{assay}{Character string. The assay from which extract the count matrix for SVD. Default "counts".}

\item{coords}{Logical. Indicates whether CA standard coordinates should be calculated. Default TRUE}

\item{princ_coords}{Integer. Number indicating whether principal coordinates should be calculated for the rows (=1), columns (=2), both (=3) or none (=0).
Default 1.}

\item{python}{A logical value indicating whether to use singular-value decomposition from the python package torch.
This implementation dramatically speeds up computation compared to `svd()` in R.}

\item{dims}{Integer. Number of CA dimensions to retain. Default NULL (keeps all dimensions).}

\item{top}{Integer. Number of most variable rows to retain. Default NULL.}

\item{inertia}{Logical.. Whether total, row and column inertias should be calculated and returned. Default TRUE.}

\item{rm_zeros}{Logical. Whether rows containing only 0s should be removed. Keeping zero only rows might lead to unexpected results.Default TRUE.}

\item{return_input}{Logical. If TRUE returns the input (SingleCellExperiment object) with the CA saved in the reducedDim slot "CA". Otherwise returns a "cacomp". Default FALSE.}
}
\value{
If return_input is FALSE returns a cacomp object with components
U, V and D: The results from the SVD.
row_masses and col_masses: Row and columns masses.
top_rows: How many of the most variable rows were retained for the analysis.
tot_inertia, row_inertia and col_inertia: Only if inertia = TRUE. Total, row and column inertia respectively.
If return_input is TRUE returns a SingleCellExperiment object with a matrix of standard coordinates of the columns in
reducedDim(obj, "CA"). Additionally, the matrix contains the following attributes:
"prin_coords_rows": Principal coordinates of the rows.
"singval": Singular values. For the explained inertia of each principal axis calculate singval^2.
"percInertia": Percent explained inertia of each principal axis.

To recompute a regular "cacomp" object without rerunning cacomp use `ca_from_sce()`.
}
\description{
`cacomp.SingleCellExperiment` performs correspondence analysis on a matrix and stores the standard coordinates
 of the columns (= cells) and the principal coordinates of the rows (= genes) as a matrix in the SingleCellExperiment container.
}
\details{
The calculation is performed according to Greenacre. Singular value decomposition
can be performed either with the base R function `svd` or preferably by the much faster
pytorch implementation (python = TRUE, requires a working python installation with numpy and pytorch). When working on large matrices, CA coordinates and
principal coordinates should only computed when needed to save computational time.
}
\references{
Greenacre, M. Correspondence Analysis in Practice, Third Edition, 2017.
}
