% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{check_cacomp}
\alias{check_cacomp}
\title{Check if cacomp object was correctly created.}
\usage{
check_cacomp(object)
}
\arguments{
\item{object}{A cacomp object.}
}
\description{
Checks if the slots in a cacomp object are of the correct size
and whether they are coherent.
}
