% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{check_names}
\alias{check_names}
\title{Check names of potential cacomp object.}
\usage{
check_names(x, canames)
}
\arguments{
\item{x}{An object to check.}

\item{canames}{Allowed names.}
}
\description{
Check names of potential cacomp object.
}
