% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{comp_ft_residuals}
\alias{comp_ft_residuals}
\title{Compute Freeman-Tukey residuals}
\usage{
comp_ft_residuals(mat)
}
\arguments{
\item{mat}{A numerical matrix or coercible to one by `as.matrix()`.
Should have row and column names.}
}
\value{
A named list. The elements are:
\itemize{
\item "S": standardized residual matrix.
\item "tot": grand total of the original matrix.
\item "rowm": row masses.
\item "colm": column masses.
}
}
\description{
TODO
}
