% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{comp_NB_residuals}
\alias{comp_NB_residuals}
\title{Compute Negative-Binomial residuals}
\usage{
comp_NB_residuals(mat, theta = 100, clip = TRUE, cutoff = NULL, freq = TRUE)
}
\arguments{
\item{mat}{A numerical matrix or coercible to one by `as.matrix()`.
Should have row and column names.}

\item{clip}{logical. Whether residuals should be clipped if they are 
higher/lower than a specified cutoff}

\item{cutoff}{numeric. Residuals that are larger than cutoff or lower than
-cutoff are clipped to cutoff.}
}
\value{
A named list. The elements are:
\itemize{
\item "S": standardized residual matrix.
\item "tot": grand total of the original matrix.
\item "rowm": row masses.
\item "colm": column masses.
}
}
\description{
Compute Negative-Binomial residuals
}
