% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{fill_names}
\alias{fill_names}
\title{Create missing names}
\usage{
fill_names(x, canames)
}
\arguments{
\item{x}{Object to fill names in for.}

\item{canames}{Allowed names.}
}
\description{
Adds missing names to object and sets them NULL.
}
