\name{first}
\alias{first}
\title{
First element in a vector   
}
\description{
Returns the first element in a vector.
}
\usage{
first(x)
}
\arguments{
  \item{x}{a vector}
}
\details{
This function takes a vector and returns only the first element.
}
\references{
G. Helzer. (1989):
\emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu}
}
\seealso{
    \code{\link[apl]{last}}, 
    \code{\link[apl]{butFirst}}, 
    \code{\link[apl]{butLast}}
}
\examples{
    x <- 1:10
    first(x)
}
\keyword{manip}

