% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex}
\alias{gtex}
\title{Subset of GTEx tissue dataset.}
\format{
A matrix with 56202 rows and 220 samples:
}
\source{
\url{https://gtexportal.org/home/datasets}
}
\usage{
gtex
}
\description{
A count matrix consisting of a subset of the
Genotype-Tissue Expression (GTEx) data.
The matrix contains count data from 22 tissues:
`unique(gsub("[[:digit:]]", "", colnames(gtex)))`
}
\keyword{datasets}
