% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{new_cacomp}
\alias{new_cacomp}
\title{Create new object of class "cacomp"}
\usage{
new_cacomp(x)
}
\arguments{
\item{x}{object to create new cacomp from.}
}
\description{
Create new object of class "cacomp"
}
