% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Generates plot for results from apl_topGO}
\usage{
plot_enrichment(genenr, ntop = 10)
}
\arguments{
\item{genenr}{data.frame. gene enrichment results table.}

\item{ntop}{numeric. Number of elements to plot.}
}
\value{
Returns a ggplot plot.
}
\description{
Generates plot for results from apl_topGO
}
