% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{print.cacomp}
\alias{print.cacomp}
\title{Prints cacomp object}
\usage{
\method{print}{cacomp}(x, ...)
}
\arguments{
\item{x}{cacomp object to print}

\item{...}{ignored.}
}
\description{
Provides more user friendly printing of cacomp objects.
}
