% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{recompute}
\alias{recompute}
\title{Recompute missing values of cacomp object.}
\usage{
recompute(caobj, mat)
}
\arguments{
\item{caobj}{A cacomp object with std_coords_cols, the prin_coords_rows and D.}

\item{mat}{A matrix from which the cacomp object derives from.}
}
\value{
A cacomp object with additional calculated row_masses, col_masses, std_coords_rows, U and V.
}
\description{
The caobj needs to have the std_coords_cols, the prin_coords_rows and D calculated. From this the remainder will be calculated.
Future updates might extend this functionality.
}
