% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{rm_zeros}
\alias{rm_zeros}
\title{removes 0-only rows and columns in a matrix.}
\usage{
rm_zeros(obj)
}
\arguments{
\item{obj}{A matrix.}
}
\value{
Input matrix with rows & columns consisting of only 0 removed.
}
\description{
removes 0-only rows and columns in a matrix.
}
