% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{run_python_SVD}
\alias{run_python_SVD}
\title{Run pytorch SVD via basilisk}
\usage{
run_python_SVD(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
list with SVD results.
}
\description{
Run pytorch SVD via basilisk
}
