% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl.R
\name{runAPL}
\alias{runAPL}
\alias{runAPL.default}
\alias{runAPL.matrix}
\alias{runAPL.SingleCellExperiment}
\alias{runAPL.Seurat}
\title{Compute and plot Association plot}
\usage{
runAPL(
  obj,
  group,
  caobj = NULL,
  dims = NULL,
  nrow = 10,
  top = 5000,
  score = TRUE,
  mark_rows = NULL,
  reps = 3,
  python = TRUE,
  row_labs = TRUE,
  col_labs = TRUE,
  ...
)

\method{runAPL}{default}(
  obj,
  group,
  caobj = NULL,
  dims = NULL,
  nrow = 10,
  top = 5000,
  score = TRUE,
  mark_rows = NULL,
  reps = 3,
  python = TRUE,
  row_labs = TRUE,
  col_labs = TRUE,
  ...
)

\method{runAPL}{matrix}(
  obj,
  group,
  caobj = NULL,
  dims = NULL,
  nrow = 10,
  top = 5000,
  score = TRUE,
  mark_rows = NULL,
  reps = 3,
  python = TRUE,
  row_labs = TRUE,
  col_labs = TRUE,
  ...
)

\method{runAPL}{SingleCellExperiment}(
  obj,
  group,
  caobj = NULL,
  dims = NULL,
  nrow = 10,
  top = 5000,
  score = TRUE,
  mark_rows = NULL,
  reps = 3,
  python = TRUE,
  row_labs = TRUE,
  col_labs = TRUE,
  ...,
  assay = "counts"
)

\method{runAPL}{Seurat}(
  obj,
  group,
  caobj = NULL,
  dims = NULL,
  nrow = 10,
  top = 5000,
  score = TRUE,
  mark_rows = NULL,
  reps = 3,
  python = TRUE,
  row_labs = TRUE,
  col_labs = TRUE,
  ...,
  assay = DefaultAssay(obj)
)
}
\arguments{
\item{obj}{A numeric matrix, Seurat or SingleCellExperiment object. For sequencing a count matrix, gene expression values with genes in rows and samples/cells in columns.
Should contain row and column names.}

\item{group}{Numeric/Character. Vector of indices or column names of the columns to calculate centroid/x-axis direction.}

\item{caobj}{A "cacomp" object as outputted from `cacomp()`. If not supplied will be calculated. Default NULL.}

\item{dims}{Integer. Number of dimensions to keep. Default NULL (keeps all dimensions).}

\item{nrow}{Integer. The top nrow scored row labels will be added to the plot if score = TRUE. Default 10.}

\item{top}{Integer. Number of most variable rows to retain. Default 5000.}

\item{score}{Logical. Whether rows should be scored and ranked. Ignored when a vector is supplied to mark_rows. Default TRUE.}

\item{mark_rows}{Character vector. Names of rows that should be highlighted in the plot. If not NULL, score is ignored. Default NULL.}

\item{reps}{Integer. Number of permutations during scoring. Default 3.}

\item{python}{A logical value indicating whether to use singular-value decomposition from the python package torch.
This implementation dramatically speeds up computation compared to `svd()` in R.}

\item{row_labs}{Logical. Whether labels for rows indicated by rows_idx should be labeled with text. Default TRUE.}

\item{col_labs}{Logical. Whether labels for columns indicated by cols_idx shouls be labeled with text. Default TRUE.}

\item{...}{Arguments forwarded to methods.}

\item{assay}{Character. The assay from which extract the count matrix for SVD, e.g. "RNA" for Seurat objects or "counts"/"logcounts" for SingleCellExperiments.}
}
\value{
Association plot (plotly object).
}
\description{
Computes singular value decomposition and coordinates for the Association plot.

Computes singular value decomposition and coordinates for the Association plot from SingleCellExperiment objects with reducedDim(obj, "CA") slot (optional).

Computes singular value decomposition and coordinates for the Association plot from Seurat objects, optionally with a DimReduc Object in the "CA" slot.
}
\details{
The function is a wrapper that calls `cacomp()`, `apl_coords()`, `apl_score()` and finally `apl()` for ease of use.
The chosen defaults are most useful for genomics experiments, but for more fine grained control the functions
can be also run individually for the same results.
If score = FALSE, nrow and reps are ignored. If mark_rows is not NULL score is treated as if FALSE.
}
