% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree Plot}
\usage{
scree_plot(df)
}
\arguments{
\item{df}{A data frame with columns "dims" and "inertia".}
}
\value{
Returns a ggplot object.
}
\description{
Plots a scree plot.
}
