% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{show.cacomp}
\alias{show.cacomp}
\title{Prints cacomp object}
\usage{
show.cacomp(object)
}
\arguments{
\item{object}{cacomp object to print}
}
\description{
Provides more user friendly printing of cacomp objects.
}
