% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{[,cacomp,ANY,ANY,ANY-method}
\alias{[,cacomp,ANY,ANY,ANY-method}
\title{Subset rows and columns of a cacomp object.}
\usage{
\S4method{[}{cacomp,ANY,ANY,ANY}(x, i = NULL, j = NULL, ..., drop = FALSE)
}
\arguments{
\item{x}{cacomp object}

\item{i}{rows to subset to.}

\item{j}{columns to subset to.}

\item{...}{Furhter arguments}

\item{drop}{Whether or not to coerce to the lowest possible dimension. Should
be FALSE!}
}
\value{
Returns a cacomp object with rows and columns subsetted.
}
\description{
Subset rows and columns of a cacomp object.
}
