% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{subset_dims}
\alias{subset_dims}
\title{Subset dimensions of a caobj}
\usage{
subset_dims(caobj, dims)
}
\arguments{
\item{caobj}{A caobj.}

\item{dims}{Integer. Number of dimensions.}
}
\value{
Returns caobj.
}
\description{
Subsets the dimensions according to user input.
}
