% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_test.R
\name{testfun}
\alias{testfun}
\alias{testfun.data.frame}
\title{testfun}
\usage{
testfun(x)

\method{testfun}{data.frame}(x, objd)
}
\arguments{
\item{x}{whatev}

\item{objd}{df}
}
\description{
testfun

testfun
}
