\name{unit}
\alias{unit}
\title{
Check to see if a vector is in the range.
}
\description{
Function to check if a unit is within a range.
}
\usage{
unit(i, n)
}
arguments{
  \item{i}{a vector to check}
  \item{n}{maximum of the range}
}
\details{
This function takes a vector and checks if each element is within the range specified by 1 to n.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu}
}

\seealso{
    \code{\link[apl]{first}}, 
    \code{\link[apl]{last}}, 
    \code{\link[apl]{butFirst}},
    \code{\link[apl]{butLast}}
}

\examples{
unit(2,10)
unit(11,10)
}
\keyword{manip}