% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{var_rows}
\alias{var_rows}
\title{Find most variable rows}
\usage{
var_rows(mat, top = 5000)
}
\arguments{
\item{mat}{A numeric matrix. For sequencing a count matrix, gene expression values with genes in rows and samples/cells in columns.
Should contain row and column names.}

\item{top}{Integer. Number of most variable rows to retain. Default 5000.}
}
\value{
Returns a matrix, which consists of the top variable rows of mat.
}
\description{
Calculates the variance of the chi-square components matrix and selects the top e.g 5000 rows.
}
