\name{apl-package}
\alias{apl-package}
\alias{apl}
\docType{package} 
\title{
APL functions in R
}
\description{
APL functions in R.
}
\details{
\tabular{ll}{
Package: \tab apl\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2013-09-1\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
APL functions in R.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu},
Masanao Yajima \email{yajima@stat.ucla.edu}

Maintainer: Masanao Yajima<yajima@stat.ucla.edu>
}
\references{
G. Helzer. (1989):
\emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
}
\keyword{ package }

%\examples{
%~~ simple examples of the most important functions ~~
%}
