% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl.R
\name{apl}
\alias{apl}
\title{Association Plot}
\usage{
apl(
  caobj,
  type = "ggplot",
  rowlabels = TRUE,
  collabels = TRUE,
  rows_idx = 1:nrow(caobj$apl_rows),
  cols_idx = caobj$group,
  row_labs = TRUE,
  col_labs = TRUE
)
}
\arguments{
\item{caobj}{An object of class "cacomp" and "APL" with apl coordinates calculated.}

\item{type}{"ggplot"/"plotly". For a static plot a string "ggplot", for an interactive plot "plotly". Default "ggplot".}

\item{rowlabels}{TRUE/FALSE. Logical whether rows should be labelled. Default TRUE.}

\item{collabels}{TRUE/FALSE. Logical whether columns should be labelled. Default TRUE.}

\item{rows_idx}{numeric vector. Indices of the rows that should be labelled. Default all rows: 1:nrow(caobj$apl_rows).}

\item{cols_idx}{numeric vector. Indices of the columns that should be labelled. Default is only to label columns making up the centroid: caobj$group.}

\item{row_labs}{Logical. Whether labels for rows indicated by rows_idx should be labeled with text. Default TRUE.}

\item{col_labs}{Logical. Whether labels for columns indicated by cols_idx shouls be labeled with text. Default TRUE.}
}
\value{
Either a ggplot or plotly object.
}
\description{
Plot an Association plot for the chosen columns.
}
\details{
For an interactive plot type="plotly" can be chosen, otherwise a static plot will returned.
The row and column coordinates have to be already calculated by `apl_coords()`.
}
