% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.cacomp}
\alias{as.cacomp}
\alias{as.cacomp,cacomp-method}
\alias{as.cacomp,Seurat-method}
\alias{as.cacomp,SingleCellExperiment-method}
\title{Create cacomp object from Seurat/SingleCellExperiment container}
\usage{
as.cacomp(obj, ...)

\S4method{as.cacomp}{cacomp}(obj, ...)

\S4method{as.cacomp}{Seurat}(obj, ..., assay = "RNA", slot = "counts")

\S4method{as.cacomp}{SingleCellExperiment}(obj, ..., assay = "counts")
}
\arguments{
\item{obj}{An object of class "Seurat" or "SingleCellExperiment"
with a dim. reduction named "CA" saved. For obj "cacomp" input is returned.}

\item{...}{Further arguments.}

\item{assay}{Character. The assay from which extract the count matrix,
e.g. "RNA" for Seurat objects or "counts"/"logcounts" for SingleCellExperiments.}

\item{slot}{character. Slot of the Seurat assay to use. Default "counts".}
}
\value{
A cacomp object.
}
\description{
Converts the values stored in the Seurat/SingleCellExperiment dimensional reduction slot "CA" to a cacomp object.
If recompute = TRUE additional parameters are recomputed from the saved values without rerunning SVD (need to specify assay to work).

as.cacomp.cacomp returns input without any calculations.

as.cacomp.Seurat: Converts the values stored in the Seurat DimReduc slot "CA" to an cacomp object.

as.cacomp.SingleCellExperiment: Converts the values stored in the SingleCellExperiment reducedDim slot "CA" to an cacomp object.
}
\details{
By default extracts std_coords_cols, D, prin_coords_rows, top_rows and dims from obj and outputs a cacomp object.
If recompute = TRUE the following are additionally recalculated (doesn't run SVD):
U, V, std_coords_rows, row_masses, col_masses.
}
