% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{clip_residuals}
\alias{clip_residuals}
\title{Perform clipping of residuals}
\usage{
clip_residuals(S, cutoff = sqrt(ncol(S)))
}
\arguments{
\item{S}{Matrix of residuals.}

\item{cutoff}{Value above/below which clipping should happen.}
}
\value{
Clipped residuals
}
\description{
Perform clipping of residuals
}
