% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{inertia_rows}
\alias{inertia_rows}
\title{Find most variable rows}
\usage{
inertia_rows(mat, top = 5000)
}
\value{
Returns a matrix, which consists of the top variable rows of mat.
}
\description{
Calculates the contributing inertia of each row which is defined as sum of squares of pearson residuals and selects the 
rows with the largested inertias, e.g. 5,000.
}
