\name{last}
\alias{last}
\title{
Last element in a vector   
}
\description{
Returns the last element in a vector.
}
\usage{
last(x)
}
\arguments{
  \item{x}{a vector}
}
\details{
This function takes a vector and returns only the last element.
}
\references{
G. Helzer. (1989):
\emph{An Encyclopedia of APL}, second edition, I-APL LTD, St. Albans, G.B.
}
\author{
  Jan de Leeuw \email{deleeuw@stat.ucla.edu}
}
\seealso{
    \code{\link[apl]{first}}, 
    \code{\link[apl]{butFirst}}, 
    \code{\link[apl]{butLast}}
}
\examples{
    x <- 1:10
    last(x)
}
\keyword{manip}