% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{new_cacomp}
\alias{new_cacomp}
\title{Create new "cacomp" object.}
\usage{
new_cacomp(...)
}
\arguments{
\item{...}{Should contain arguments forwared to new().
Arguments should be the slot name followed by the approbiate object type.}
}
\description{
Creates new cacomp object.
}
